+++ CHANGELOG.rst
@@ -1,6 +1,13 @@
 Changelog
 =========
 
+Release *v1.1.3* - ``2020-05-01``
+---------------------------------
+* Fix failure when using ``internal_version``
+* Add a way to chose compression level
+* Add support for Django 2.2 and 3.0
+* Add support for Python 3.8
+
 Release *v1.1.2* - ``2018-11-27``
 ---------------------------------
 * Add support for Django 2.1